Initial steps in removing multipart special treatment
Former-commit-id: 247a813ee05feecba14d692e26a1681a72920961

+++ block/s3/stream.go
@@ -15,20 +15,20 @@ const (
 	StreamingContentEncoding = "aws-chunked"
 )
 
-func CalculateStreamSizeForPayload(sizeBytes int, chunkSize int) int {
+func CalculateStreamSizeForPayload(sizeBytes int64, chunkSize int) int64 {
 	// for each chunk we need 81 fixed chars, i.e. ;chunk-signature=0055627c9e194cb4542bae2aa5492e3c1575bbb81b612b7d234b86a503ef5497
 	// we also add a constant \r\n after the sig, and another one after the payload, for a total for 85 fixed chars.
 	// the actual boundary is prefixed by the hexadecimal string size representation of the chunk which can vary in size
 	// so we need to calculate it
 	const chunkSigSize = 85
-	fullChunks := sizeBytes / chunkSize
-	partialChunkSize := sizeBytes % chunkSize
+	fullChunks := sizeBytes / int64(chunkSize)
+	partialChunkSize := sizeBytes % int64(chunkSize)
 
-	overhead := fullChunks * (chunkSigSize + len(fmt.Sprintf("%x", chunkSize))) // one sig for every chunk
+	overhead := fullChunks * (chunkSigSize + int64(len(fmt.Sprintf("%x", chunkSize)))) // one sig for every chunk
 	if partialChunkSize > 0 {
-		overhead += chunkSigSize + len(fmt.Sprintf("%x", partialChunkSize)) // last partial chunk if there is one
+		overhead += chunkSigSize + int64(len(fmt.Sprintf("%x", partialChunkSize))) // last partial chunk if there is one
 	}
-	overhead += chunkSigSize + len(fmt.Sprintf("%x", 0)) // trailing sig
+	overhead += chunkSigSize + int64(len(fmt.Sprintf("%x", 0))) // trailing sig
 	return sizeBytes + overhead
 }
